add and remove images from item editor
