+++ static/scripts/itemEditor.js
@@ -269,21 +269,6 @@ async function addTab(type, name, force=false) {
           ] })
         )),
       ] }),
-      // createElement('button', { attrs: {
-      //   type: 'button',
-      //   innerText: 'Add New Image',
-      //   onclick: () => {
-      //     const url = getIdValue('new_gallery_image');
-      //     const label = getIdValue('new_gallery_image_label');
-      //     if (!url) return;
-      //     const newState = { ...obj_data };
-      //     if (!('imgs' in newState.gallery)) newState.gallery.imgs = [];
-      //     newState.gallery.imgs.push({ url, label });
-      //     updateObjData(newState);
-      //     resetTabs(name);
-      //   },
-      // } }),
-      // createElement('input', { attrs: { id: 'new_gallery_image', placeholder: T('Image URL') } }),
       createElement('button', { attrs: {
         type: 'button',
         innerText: 'Upload Image',
@@ -294,7 +279,7 @@ async function addTab(type, name, force=false) {
             if (!url) return;
             const newState = { ...obj_data };
             if (!('imgs' in newState.gallery)) newState.gallery.imgs = [];
-            newState.gallery.imgs.push({ url, label, name: newName });
+            newState.gallery.imgs.push({ id: newId, url, label, name: newName });
             updateObjData(newState);
             resetTabs(name);
           });
